+++ lib/convert.dart
@@ -7,7 +7,7 @@ library convert;
 export 'src/accumulator_sink.dart';
 export 'src/byte_accumulator_sink.dart';
 export 'src/codepage.dart';
-export 'src/fixed_datetime_codec.dart';
+export 'src/fixed_datetime_formatter.dart';
 export 'src/hex.dart';
 export 'src/identity_codec.dart';
 export 'src/percent.dart';
