+++ test/fixed_datetime_formatter_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:convert/src/fixed_datetime_codec.dart';
+import 'package:convert/src/fixed_datetime_formatter.dart';
 import 'package:test/test.dart';
 
 void main() {
